+++ src/pl/plajer/villagedefense3/arena/ArenaRegistry.java
@@ -83,12 +83,12 @@ public class ArenaRegistry {
   }
 
   public static void registerArena(Arena arena) {
-    Main.debug("Registering new game instance, " + arena.getID(), System.currentTimeMillis());
+    Main.debug(Main.LogLevel.INFO, "Registering new game instance, " + arena.getID());
     arenas.add(arena);
   }
 
   public static void unregisterArena(Arena arena) {
-    Main.debug("Unegistering game instance, " + arena.getID(), System.currentTimeMillis());
+    Main.debug(Main.LogLevel.INFO, "Unegistering game instance, " + arena.getID());
     arenas.remove(arena);
   }
 
@@ -110,7 +110,7 @@ public class ArenaRegistry {
   }
 
   public static void registerArenas() {
-    Main.debug("Initial arenas registration", System.currentTimeMillis());
+    Main.debug(Main.LogLevel.INFO, "Initial arenas registration");
     if (ArenaRegistry.getArenas() != null) {
       if (ArenaRegistry.getArenas().size() > 0) {
         for (Arena arena : ArenaRegistry.getArenas()) {
@@ -187,7 +187,7 @@ public class ArenaRegistry {
       Bukkit.getConsoleSender().sendMessage(ChatManager.colorMessage("Validator.Instance-Started").replace("%arena%", ID));
     }
     new ShopManager();
-    Main.debug("Arenas registration completed", System.currentTimeMillis());
+    Main.debug(Main.LogLevel.INFO, "Arenas registration completed");
   }
 
   public static List<Arena> getArenas() {