+++ src/pl/plajer/villagedefense3/commands/GameCommands.java
@@ -154,11 +154,11 @@ public class GameCommands extends MainCommand {
       p.sendMessage(ChatManager.PLUGIN_PREFIX + ChatManager.colorMessage("Commands.Teleported-To-The-Lobby"));
       if (plugin.isBungeeActivated()) {
         plugin.getBungeeManager().connectToHub(p);
-        System.out.print(p.getName() + " is teleported to the Hub Server");
+        Main.debug(Main.LogLevel.INFO, p.getName() + " was teleported to the Hub server");
       } else {
         ArenaRegistry.getArena(p).teleportToEndLocation(p);
         ArenaManager.leaveAttempt(p, ArenaRegistry.getArena(p));
-        System.out.print(p.getName() + " has left the arena! He is teleported to the end location.");
+        Main.debug(Main.LogLevel.INFO, p.getName() + " has left the arena! He is teleported to the end location.");
       }
     }
   }